requirements.txt
google-genai
Pillow
reportlab
